remove bitmask returns from createMOEWAsync() and update testing to solely test createMOEWAsync() and the bitmask calculation utils function separately
